% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiD.R
\name{DiD}
\alias{DiD}
\title{non-param diff-in-diff}
\usage{
DiD(X, Y, Ti, Si, constant_eff = c("constant", "non_constant"),
  gamma = NULL, k_folds = 10, lambda_choice = "lambda.min",
  penalty_factor = NULL)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed response (real valued)}

\item{Ti}{the time variable (0 or 1)}

\item{Si}{the state variable (0 or 1)}

\item{constant_eff}{whether the treatment effect is modeled to be constant, choose from "constant" or "non_constant"
if assuming a constant treatment effect, set constant_eff = "constant" (which runs the Transformed Regression from the paper);
if assuming a non-constant effect, set constant_eff = "non_constant"}

\item{gamma}{balancing weights, a vector of the same length as the length of Y (observables),
used for balancing in the treatment estimation; can be left blank, in which case the algorithm
does not use balancing}

\item{k_folds}{number of folds for cross-validation; default is 10}

\item{lambda_choice}{choose how to cross-validate; choose from "lambda.min" or "lambda.1se"}

\item{penalty_factor}{user specified penalty factor, used in the R learner}
}
\value{
a list including TAU_hat, estimated ATE; std.err.est, estimated standard error;
tau_hat, non-parametric point estimates of treatment effect; y.est, estimated outcome values.
}
\description{
Non-parametric diff-in-diff estimation, as proposed by Lu, Nie, Wager (2019),
 implemented with the lasso (glmnet) and the R learner (rlearner) from Nie and Wager (2017)
}
