% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.basis.R
\name{generate.basis}
\alias{generate.basis}
\title{generate hermite basis}
\usage{
generate.basis(X, order = 3)
}
\arguments{
\item{X}{an n by p matrix containing the input features}

\item{order}{order of hermite polynomials that we will use}
}
\value{
a n by q matrix, which is a basis expansion of the input X
}
\description{
Generates hermite basis for a given order, which we use for
non-parametric regression
}
