% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimax.R
\name{minimax}
\alias{minimax}
\title{Minimax balancing weights}
\usage{
minimax(X, Ti, Si, zeta = 0.5, solver = c("ECOS", "SCS"),
  verbose = FALSE)
}
\arguments{
\item{X}{the input features}

\item{Ti}{the time variable (0 or 1)}

\item{Si}{the state variable (0 or 1)}

\item{zeta}{tuning parameter, in [0,1]}

\item{solver}{cvx opmitization solver, choose from "ECOS" or "SCS"}
}
\value{
a vector gamma of the same length as Ti
}
\description{
function that finds the minimax balancing weights, using the method in
Lu, Nie, Wager (2019)
}
