% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_learner.R
\name{naive_learner}
\alias{naive_learner}
\title{baseline methods used in simulations}
\usage{
naive_learner(X, Y, Ti, Si, constant_eff = c("constant", "non_constant"))
}
\arguments{
\item{X}{input features}

\item{Ti}{the time variables (0 or 1)}

\item{Si}{the state variables (0 or 1)}

\item{constant_eff}{whether we are assuming a constant or non-constant
treatment effect (choose from "constant" and "non_constant")}
}
\value{
estimated ATE and standard error
}
\description{
two simple baselines: if the treatment effect is assumed to be
constant, then we estimate ATE by running OLS; if the treatment effect is not
assumed to be constant, then ATE is estimated by taking sample averages of
the outcomes Y (details see Lu, Nie, Wager (2019))
}
