% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{plugin}
\alias{plugin}
\title{Plugin weights}
\usage{
plugin(X, Ti, Si, k_folds = 10)
}
\arguments{
\item{X}{input features}

\item{Ti}{time variables}

\item{Si}{state variables}

\item{k_folds}{number of folds for cross-fitting}
}
\value{
a vector gamma.plugin of weights of the same length as Ti
}
\description{
finds balancing weights using the plugin estimator for P(Ti=1 | X) etc;
plugin weights estimated using lasso (glmnet); weights are used for methods such as
IPW and DiD-AIPW
}
